Consolidate network.utils and network.util
Remove unused utility methods previously using with networkx.

[#118745557]
